+++ lib/groupdate/series.rb
@@ -56,7 +56,12 @@ module Groupdate
           lambda{|k| (k.is_a?(String) ? utc.parse(k) : k.to_time).in_time_zone(@time_zone) }
         end
 
-      count = Hash[ relation.send(method, *args, &block).map{|k, v| [multiple_groups ? k[0...@group_index] + [cast_method.call(k[@group_index])] + k[(@group_index + 1)..-1] : cast_method.call(k), v] } ]
+      count =
+        begin
+          Hash[ relation.send(method, *args, &block).map{|k, v| [multiple_groups ? k[0...@group_index] + [cast_method.call(k[@group_index])] + k[(@group_index + 1)..-1] : cast_method.call(k), v] } ]
+        rescue NoMethodError
+          raise "Be sure to install time zone support - https://github.com/ankane/groupdate#for-mysql"
+        end
 
       series =
         case @field